+++ friendlysam/sequences.py
@@ -1,5 +1,8 @@
 # -*- coding: utf-8 -*-
 
+import logging
+logger = logging.getLogger(__name__)
+
 class Item(object):
     """docstring for Item"""
     def __init__(self, sequence, value):