+++ friendlysam/solvers/__init__.py
@@ -1,1 +1,4 @@
 # -*- coding: utf-8 -*-
+
+import logging
+logger = logging.getLogger(__name__)